requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
xlrd>=2.0.0
pypdf>=3.0.0
reportlab>=4.0.0
PyMuPDF>=1.23.0
PyYAML>=6.0
